@@ -1,23 +1,55 @@
 apply plugin: 'com.android.application'
 
 android {
-    compileSdkVersion 25
-    buildToolsVersion "25.0.3"
+    compileSdkVersion 31
+    buildToolsVersion "30.0.3"
     defaultConfig {
-        applicationId "com.dw.mq.ysou"
+        applicationId "com.easou.espay_user_h5"
         minSdkVersion 16
-        targetSdkVersion 25
-        versionCode 1
-        versionName "1.0"
-        testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
+        targetSdkVersion 31
+        versionCode 4
+        versionName "4.0"
+        vectorDrawables.useSupportLibrary =true
+        testInstrumentationRunner 'androidx.test.runner.AndroidJUnitRunner'
     }
+    compileOptions {
+        sourceCompatibility JavaVersion.VERSION_1_8
+        targetCompatibility JavaVersion.VERSION_1_8
+    }
+
+    //签名配置
+    signingConfigs {
+        release {
+            keyAlias RELEASE_KEY_ALIAS
+            keyPassword RELEASE_KEY_PASSWORD
+            storeFile file(RELEASE_STORE_FILE)
+            storePassword RELEASE_STORE_PASSWORD
+        }
+        debug {
+            storeFile file("debug.keystore")
+            keyAlias "AndroidDebugKey"
+            keyPassword "android"
+            storePassword "android"
+        }
+    }
+
     buildTypes {
         release {
             minifyEnabled true
+            zipAlignEnabled true
             proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
         }
+
+        debug {
+            minifyEnabled false
+            zipAlignEnabled true
+            proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
+            signingConfig signingConfigs.release
+        }
     }
-    configurations.all {
+
+
+  /*  configurations.all {
         resolutionStrategy.eachDependency { DependencyResolveDetails details ->
             def requested = details.requested
             if (requested.group == 'com.android.support') {
@@ -26,18 +58,20 @@ android {
                 }
             }
         }
-    }
+    }*/
 }
 
 dependencies {
     compile fileTree(include: ['*.jar'], dir: 'libs')
-    androidTestCompile('com.android.support.test.espresso:espresso-core:2.2.2', {
+    androidTestCompile('androidx.test.espresso:espresso-core:3.1.0', {
         exclude group: 'com.android.support', module: 'support-annotations'
     })
-    compile 'com.android.support:appcompat-v7:25.+'
+    compile 'androidx.appcompat:appcompat:1.0.0'
 //    compile 'com.android.support:appcompat-v7:28.0.0'
-    compile 'com.android.support.constraint:constraint-layout:1.0.2'
+    compile 'androidx.constraintlayout:constraintlayout:1.1.3'
 
     testCompile 'junit:junit:4.12'
     compile project(':espay_user_lib')
+    implementation 'com.google.android.gms:play-services-auth:20.3.0'
+    implementation "com.android.billingclient:billing:4.1.0"
 }
