@@ -1,16 +1,16 @@
 apply plugin: 'com.android.library'
 
 android {
-    compileSdkVersion 25
-    buildToolsVersion "25.0.3"
+    compileSdkVersion 31
+    buildToolsVersion "30.0.3"
 
     defaultConfig {
         minSdkVersion 16
-        targetSdkVersion 25
+        targetSdkVersion 31
         versionCode 1
         versionName "1.0"
         consumerProguardFiles 'proguard-rules.pro'
-        testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
+        testInstrumentationRunner 'androidx.test.runner.AndroidJUnitRunner'
     }
     buildTypes {
         release {
@@ -23,7 +23,7 @@ android {
         }
     }
 
-    configurations.all {
+   /* configurations.all {
         resolutionStrategy.eachDependency { DependencyResolveDetails details ->
             def requested = details.requested
             if (requested.group == 'com.android.support') {
@@ -32,29 +32,21 @@ android {
                 }
             }
         }
-    }
+    }*/
 }
 
 dependencies {
     compile fileTree(dir: 'libs', include: ['*.jar'])
-    androidTestCompile('com.android.support.test.espresso:espresso-core:2.2.2', {
+    androidTestCompile('androidx.test.espresso:espresso-core:3.1.0', {
         exclude group: 'com.android.support', module: 'support-annotations'
     })
 //    compile 'com.android.support:appcompat-v7:28.0.0'
-    compile 'com.android.support:appcompat-v7:25.+'
+    compile 'androidx.appcompat:appcompat:1.0.0'
     testCompile 'junit:junit:4.12'
-    compile files('libs/alipaySdk-15.8.02.jar')
     compile files('libs/gson-1.5.jar')
-    compile files('libs/HeepaySDK_4.0.5.jar')
-    compile files('libs/locSDK_4.2.jar')
-    compile files('libs/payecoplugin_standard_baidu_3-2.1.12.jar')
-    compile files('libs/ulo5sdk-1.5.2.jar')
     compile files('libs/oaid_sdk_1.1.0.jar')
-    compile files('libs/RangersAppLog-Lite-cn-5.3.0.jar')
-    compile files('libs/gism_1.3.3.jar')
-    compile files('libs/GDTActionSDK.min.1.8.2.jar')
-    compile files('libs/monitorsdk-1.0.10.jar')
-//    compile files('libs/Baidu_ocpc_2.5.jar')
-    compile files('libs/humesdk-1.0.0.jar')
-    compile files('libs/qilintrans2.0.jar')
+    compile files('libs/AF-Android-SDK.jar')
+    implementation 'com.google.android.gms:play-services-auth:20.3.0'
+    implementation "com.android.billingclient:billing:4.1.0"
+    implementation 'com.facebook.android:facebook-login:latest.release'
 }
